+++ CHANGELOG.md
@@ -23,6 +23,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Donor Profiles are now extensible (#5577)
 
 ### New
+-   Migrations table (#5580)
+-   Legacy functions used for logging are now updated to use the new Logging API (#5614)
+-   The new Logs UI is implemented (#5591)
 
 -   Add since-unreleased script for updating @unreleased docblocks. (#5602)